fikser kommentarer
